Split fixtures in the Menu module
