+++ saleor/menu/tests/fixtures/__init__.py
@@ -1,0 +1,3 @@
+from .menu import *  # noqa: F403
+from .menu_item import *  # noqa: F403
+from .menu_item_translation import *  # noqa: F403
